Progress for demo, version independent running of components

+++ cli/UpgradeComponent.java
@@ -1,27 +1,27 @@
 package grakn.core.cli;
 
-import picocli.CommandLine;
 import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Parameters;
 
-import java.io.File;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Comparator;
 import java.util.Objects;
 import java.util.concurrent.Callable;
 
 @Command(name = "upgrade", description = "Upgrades a Grakn component.")
 public class UpgradeComponent implements Callable<Integer> {
 
+    private static final String RELEASE_REPO_NAME = "distribution";
+    private static final String SNAPSHOT_REPO_NAME = "distribution-snapshot";
+
     private static final String REPO_TEMPLATE =
-            "https://repo.grakn.ai/repository/distribution-snapshot/{group-name}/{artifact-name}/{version}/{artifact-name}-{version}.tgz";
+            "https://repo.grakn.ai/repository/{repo-name}/{group-name}/{artifact-name}/{version}/{artifact-name}-{version}.tgz";
     private static final String DOWNLOAD_DIR = "downloads";
     private static final String DOWNLOAD_PATH_TEMPLATE = "{group-name}/{artifact-name}/{version}/{artifact-name}-{version}.tgz";
 
@@ -30,6 +30,9 @@ public class UpgradeComponent implements Callable<Integer> {
     public UpgradeComponent(GraknComponentDefinition[] availableComponents) {
         this.availableComponents = availableComponents;
     }
+
+    @Option(names = {"-s", "--snapshot"}, description = "Use a snapshot version from the snapshot repo.")
+    boolean snapshot;
 
     @Parameters(index = "0", description = "Component to upgrade.")
     String componentToUpgrade;
@@ -43,6 +46,9 @@ public class UpgradeComponent implements Callable<Integer> {
 
         @Option(names = {"-a", "--archive"}, description = "A local path to a tar.gz to upgrade from.")
         Path path;
+
+        @Option(names = {"--remove"}, hidden = true)
+        boolean remove;
     }
 
     @Override
@@ -56,6 +62,9 @@ public class UpgradeComponent implements Callable<Integer> {
             return 1;
         }
 
+        Path componentPath = graknHome.resolve(componentToUpgrade);
+        Path backupPath = graknHome.resolve(componentToUpgrade + "-BAK");
+
         Path downloadPath;
         if (upgradeFrom.version != null) {
             URL url = generateDownloadURL(availableComponent);
@@ -64,24 +73,25 @@ public class UpgradeComponent implements Callable<Integer> {
             System.out.println("Downloading " + url.toString() + " ...");
             GraknFileManager.download(url, downloadPath);
             System.out.println("Downloading complete.");
-        } else {
+        } else if (upgradeFrom.path != null) {
             downloadPath = upgradeFrom.path;
+        } else if (upgradeFrom.remove) {
+            GraknFileManager.recursiveDeleteIfExists(componentPath);
+            return 0;
+        } else {
+            throw new IllegalStateException();
         }
 
-        Path componentPath = graknHome.resolve(componentToUpgrade);
-        Path backupPath = graknHome.resolve(componentToUpgrade + "-BAK");
-
-        // Back-up old component
-        Files.move(componentPath, backupPath);
+        // Back-up old component if it exists
+        if (Files.exists(componentPath)) {
+            Files.move(componentPath, backupPath);
+        }
 
         // Un-targz the new component
         GraknFileManager.unTarGz(downloadPath, componentPath);
 
-        // Delete backup folder
-        Files.walk(backupPath)
-                .sorted(Comparator.reverseOrder())
-                .map(Path::toFile)
-                .forEach(File::delete);
+        // Delete backup folder if it exists
+        GraknFileManager.recursiveDeleteIfExists(backupPath);
 
         if (upgradeFrom.version != null) {
             // Delete temporary download and dirs by backwards recursion through the path
@@ -101,6 +111,7 @@ public class UpgradeComponent implements Callable<Integer> {
 
     private String replaceVars(String path, GraknComponentDefinition component) {
         return path
+                .replace("{repo-name}", snapshot ? SNAPSHOT_REPO_NAME : RELEASE_REPO_NAME)
                 .replace("{group-name}", component.getComponentDistributionGroup())
                 .replace("{artifact-name}", component.getComponentDistributionArtifact())
                 .replace("{version}", upgradeFrom.version);
